+++ modal-go/app.go
@@ -15,7 +15,8 @@ import (
 // App references a deployed Modal App.
 type App struct {
 	AppId string
-	ctx   context.Context
+	Name string
+	ctx  context.Context
 }
 
 // LookupOptions are options for finding deployed Modal objects.
@@ -118,7 +119,7 @@ func AppLookup(ctx context.Context, name string, options *LookupOptions) (*App, 
 		return nil, err
 	}
 
-	return &App{AppId: resp.GetAppId(), ctx: ctx}, nil
+	return &App{AppId: resp.GetAppId(), Name: name, ctx: ctx}, nil
 }
 
 // CreateSandbox creates a new Sandbox in the App with the specified image and options.